[mk/depends] Remove default CPPFLAGS and LDFLAGS computed by dependencies
By default, the dependency mechanism was adding -I and -L directives to the
directories of the resolved dependencies. This is actually pretty much useless
since packages usually do that themseleves. This can also be troublesome
because those flags usually end up in the beginning of the {C,LD}FLAGS and can
cause installed files to be included instead of the packages' own files.

So get rid of this default behaviour, but keep the variable semantics so that
individual packages can still use this behaviour if needed.