[meta-pkgs/gearbox] Import gearbox-9.11
Gearbox is a collection of software tools which could be integrated into
complete robotic systems with one of many software integration
frameworks. GearBox is not a framework and thus does not compete with any
existing or future frameworks. It does not contain any "glue" code to help you
assemble a complete system.

This package contains all the individual Gearbox tools and libraries that have
been packages in robotpkg.