[mk/configure] Improve patching libtool/configure scripts that don't need c++
configure+libtool scripts tend to look for a c++ preprocessor even for packages
that don't require c++. Packages using gnu configure and no c++ are patched so
that this error is not fatal.

In the case of pkgtools/pkgconfig, there is a sub-configure script in a
subdirectory, so that the robotpkg SUBST step for patching configure needs to
look one level of directory below. This is done by looking for "configure
*/configure".

Fixes https://softs.laas.fr/bugzilla/show_bug.cgi?id=125