[hardware/gyroKVH5000] Fix WRKSRC definition and a `make checkout` issue
WRKSRC must be defined to the root of the extracted source tree. Making it
point in the middle of the tree breaks the `make checkout` target.

In the case of this package, what was really meant was to redefine
CONFIGURE_DIRS (path relative to WRKSRC where the configure script resides).

While here, fix old-style dependency on devel/mkdep.