[mk] Build only for platforms known to work, unless explicitely set
If a package don't set one of {NOT,ONLY}_FOR_PLATORM, accept to build only
for Linux-% and NetBSD-%-i386. Those are the only two platforms more or less
tested.

We can probably add SunOS-%-i86pc to the default list if our official tester
on this platform requests it!