[mk] Add the possibility to do `make confirm` to force building on a platform
Doing `make confirm` or `make target confirm` (with target being a valid make
target) will skip {NOT,ONLY}_FOR_PLATFORM tests. Useful to check the actual
build status of a package on a platform disabled by default.

While here, contribute to the cleaning of mk/robotpkg.mk by moving tests that
can disable a package to the can-be-built-here.mk file.