[mk] On Darwin, filter PLIST with shlib-dylib.awk to substitute .so by .dylib
PLISTs are ELF-centric, they name the shared libraries .so (that's a feature).
To handle Darwin's dylib, the .so files in PLIST that don't exist but for which
a dylib is found are dynamically changed.

This could not be done by simply defining a SHLIB_EXT variable in the PLIST
because library names like .so.x.y.z are named .x.y.z.dylib on Darwin, so the
problem is more complex than it looks like.

As a side effect, if you build a PLIST (with print-PLIST) on Darwin, you will
have to rename .dylib to .so manually. Unless this is really (really) annoying
someone, this will probably be considered as a feature as well.