[pkgtools/robotpkgdb] Add license information in the databases
Licenses for each package are defined in the LICENSE package variable.

For the CSV database: column #16 contains the name of the license (or an empty
string).

For the SQLite database: the column pkginfo.license contains the name of the
license (or an empty string). A separate 'licences' table contains two columns
: 'name' and 'text'. Text contains the full text of the license.