[mk] Make LICENSE variable a mandatory item in a package description
Optional LICENSE variable is now mandatory. During the transition, only a
warning will be printed for packages that do not define a LICENSE.

By default, ACCEPTABLE_LICENSES contains the following list :
   2-clause-bsd			modified-bsd 	       original-bsd
   gnu-gpl-v2			gnu-gpl-v3
   gnu-lgpl-v2			gnu-lgpl-v2.1		gnu-lgpl-v3

Setting ACCEPTABLE_LICENSES in robotpkg.conf with the += operator will append
to this list. Setting ACCEPTABLE_LICENSES with the = operator will override
this default.

LICENSE shall contain the name of the package license, and match one of the
files in the toplevel licenses directory. The naming convention for licenses
should follow pkgsrc style:

(quoted from the pkgsrc documentation :
       Licenses are collected in the licenses/ subdirectory of
       pkgsrc.  For open source license, we generally use the same
       names as either the Open Source Initiative or the Free
       Software Foundation, choosing a name to reduce confusion.
       Open source and free licenses do not have a -license suffix,
       and non-free licenses do.

       http://opensource.org/licenses/category
       http://www.fsf.org/licensing/licenses/index_html
)