[licenses] Add doc license
doc license is the license for the ACE, TAO and other software from the
Washington university. Whether this license should deserve a -license suffix is
really not obvious (see 4953cdb47). On the one hand it was not approved by
OSI because of its 'redundant, not reusable' character. On the other hand, it
is mostly an Apache 2.0 license. So lets call is 'doc' and add it to the
DEFAULT_ACCEPTABLE_LICENSE because of its permissive nature.