[mk] Allow a CHECKOUT.<pkg> variable override or define a revision to checkout
A specific revision of a package can be checkout from a repository by defining
either a CHECKOUT.<pkg> variable (highest priority) or a CHECKOUT variable
(lowest priority).

CHECKOUT.<pkg> is meant to be used in a user's robotpkg.conf
CHECKOUT can be defined in a package Makefile for a permanent effect.

Both can of course be defined on the command line for sporadic use.