[architecture/genom] Fix PKG_OPTION_UNSET.api definition in depend.mk
The genom depend.mk must allow packages to set PKG_OPTION_UNSET.api and has to
add additional definitions to it. Since this variable is normally defined
verbatim, some tricky construction must be used to append to it. Typically,
using += don't work because += always adds an extra space. The only way to do
this properly is to use the $(value) function.