[motion/libmwalk] Create libmwalk-1.1
libmwalk is a walk pattern generator for the HRP-2 robots. It depends on
OpenHRP's dynamic model and model loader libraries and requires the proprietary
HRP-2 model provided by the (non-free) hrpsys-jrl package.

libmwalk C++ interface accepts a stack of footprints, specified in 6 dimensions
(position and orientation) plus the supporting leg. Each footprint is relative
to the previous position of the robot and especially that of the previous
supporting leg. The generator is reactive in the sense that all footprints in
the stack can be modified dynamically (except for the very next footprint to be
executed).