[mk] Add the 'headings' package target that generates headings for Makefiles
The package Makefiles, depend.mk and other similar configuration files cannot
be copyrighted. The 'headings' target replace the first lines of comments in
those files by a standard header mentioning the author and creation date.

If you need to add a comment in a package Makefile, separate it from the first
block of comment by at least one empty line.