[architecture/genom] Fix the option script for the '-api' option.
Make sure not to evaluate the option script defined in PKG_OPTION_UNSET.api too
early, so that variables in there are only dereferenced when they are supposed
to. This is done by using the $(value) gmake function that returns the raw
value of a variable.

Fixes the build of mapping/dtm-genom with the 'gdhe' option.