[image/sift] Add LICENSE
This package is available under a modified-bsd license. However, the methods
embodied in the software are subject to the following US patent:

"Method and apparatus for identifying scale invariant features
in an image and use of same for locating an object in an image," David
G. Lowe, US Patent 6,711,293 (March 23, 2004).

I am not sure what should be done in robotpkg regarding this patent.
While here, clean whitespace in DESCR and write a better COMMENT.