[mk/update] Clean and rewrite the 'update' target
The code for doing 'update' was overly obfuscated in my opinion, probably to
handle some obscure corner cases in pkgsrc. In robotpkg, it seems that we do
not need all those corner cases, so I simplified everything in the hope that it
will be more robust - at least easier to track down potential problems.

The behaviour did not change: when UPDATE_TARGET is undefined, a sane value is
computed. It default to deinstalling all packages depending on the one being
updated, then updating the package, then reinstalling all deinstalled packages.
When UPDATE_TARGET is set to 'replace', all the packages are replaced one by
one with 'make replace'.

The code better handle the situation where an update is resumed, to avoid
redoing things already done (but I might have introduced new bugs for this
situation).