[mk/plist] Change the barrier type of print-PLIST
print-PLIST might trigger an install if the package has to been installed
already. In this case, the full dependencies of the package are required, so
print-PLIST needs to run after the 'depends' barrier instead of the simpler
'bootstrap-depends' barrier. Doing so sytematically may also help for the PLIST
variables substitution, if they depends on variables set by the depends target.