[mk/fetch] Fix definitions of optional arguments for the tnftp program
Each fetch program can define a set of arguments passed to the fetch script.
These options are defined by variables ending in .<cmd>, where <cmd> is the
basename of the fetch program. robotpkg defines defaults for the tnftp program
that is normally used, but the suffix of the option variable names was wrong
(not updated since pkgsrc import). Thus, replaced .ftp by .tnftp.

In particular, this allows to pass correctly the -o option to name the output
file according to what is in the DISTFILES. This is especially important when
DISTFILES contains names with whitespace (you have to use URL escape codes for
the DISTNAME in this case, though).