[mk/fetch] Use FETCH_OUTPUT_ARG to rename downloaded files systematically.
The FETCH_OUTPUT_ARG used to rename the downloaded file was used only in case
of a transfer resume. I see no reason why it could not be used systematically,
so use it also in the case of a regular download.

This fixes an issue when a DISTFILE has URL escape codes in it, because tnftp
normaly unescape those characters for naming the downloaded file, which is
something we want to avoid in robotpkg.