[mk/sets] Implement packages sets, for defining collection of packages.
Instead of installing, removing or updating packages one-by-one, you can define
collections of packages in robotpkg.conf.  Once one or more collections are
defined, they enable special targets that work on all the packages of a
collection.

To define a collection, you have to give it a name and list the set of packages
forming the collection in the special PKGSET.<name> variable in robotpkg.conf
and set it to a <list>, where <name> is the name of the collection (any string
is valid) and <list> is the list of packages in the collection, in the form
<category>/<name>. For instance,
   PKGSET.myset = architecture/genom devel/pocolibs
defines  a collection named  myset  that contains  the two  packages genom and
pocolibs.

For each collection <name>, the following targets are available: clean-<name>,
clean-depends-<name>, fetch-<name>, extract-<name>, install-<name>,
replace-<name>, update-<name>. They perform the same action as their respective
counterpart without -<name> suffix, expect that they work on all packages of
the set. In addition, for the replace and update targets, they sort the
packages in the order of their dependencies so that the job is done a sensible
order.

For  the user  convenience,  A  special   installed  collection is  always
defined  and represents all  currently installed  packages. Thus,  invoking the
update-installed target will update all currently installed packages.

Two  robotpkg.conf variables affect the behaviour of  robotpkg
regarding packages sets:

PKGSET_FAILSAFE: When working on a set, and this variable is set to
   yes, robotpkg will continue with further packages instead of stopping on an
   error. If set to 'no', stop on first error. Default: no.

PKGSET_STRICT: Specify if package sets should be considered as 'strict' or
   include dependencies of packages defined in the set. If set to 'yes', only
   package strictly defined in sets are considered. If set to 'no',
   dependencies of packages listed in sets are added to their respective
   sets. Default: no.