[mk] Make 'replace' and 'update' a noop if the package is already up-to-date
A package is considered up-to-date if it installed, with the version defined
its Makefile, with no unsafe_depends or unsafe_depends_strict tags.
In this case, make replace and make update will not attempt to do anything.

The package can be forced to be replaced or updated by adding 'confirm' to the
command line.