[hardware/libmesasr] Avoid mirroring distfiles on robotpkg backup site.
I am not sure if we can mirror the SR driver, so in doubt set the flags
NO_PUBLIC_SRC to avoid mirroring the distfile on the backup site and
NO_PUBLIC_BIN if we ever go to binary packages.

While here, avoid useless dependency on libtool and pkg-config, and simplify
the VERSION definition. Also removed debugging do-foo target.