[mk/install] Add missing barrier definition in install target
Before 44e95238db, the missing include file was implicitely triggered by the
'replace' target. The rewrite of 'replace' dependencies triggered the latent
bug.

While at this, add missing barrier requirement for 'replace' and 'update'
targets and re-arrange those targets dependencies so that we only test when
needed if the package is already up-to-date (avoid spawning useless processes).