[devel/hpp-template-corba] Release hpp-template-corba 0.4 (initial release).
This package is intended to ease construction of CORBA servers by
templating actions that are common to all servers.

This package only contains one template class hpp::corba::Server. The
parameter of this template class is the implementation of an idl
interface.

For more information, install the package and see the html documentation.