[hardware/spnav-kmod] Delete package
This package was not the official spacenav package as found on sourceforge, and
it was only used by old versions of mouse3d-genom. Since mouse3d-genom is now
standalone, remove this useless package.

Moreover, recent kernels (linux or others) are able to read the 3dconnexion
devices without any custom driver, making obsolete the need for a specific
kernel driver.