[mk/internal] print-summary-data filters out bootstrap depends.
The build_depends output by print-summary-data do not include anymore the
'bootstrap' kind of dependencies. This changes in particular the index.html and
the robotpkgdb information.

The rationale is that 'bootstrap' dependencies are those which are necessary
only for internal robotpkg machinery, and is not considered relevant
information outside robotpkg's world.