[mk/update] Change the update target from 'reinstall' to 'install'
Update was changed some time ago to do 'reinstall' instead of 'install' in
order to better deal with stalled working directories. This was a bad idea
because this broke 'update' when resuming a previous failed update (reinstall
will fail for packages already reinstalled, unlike 'install' that will simply
do nothing). So switch back to calling 'install'.