[pkgtools/install-sh] Fix depend.mk so that it doesn't export INSTALL too early
The depend.mk exports INSTALL and related programs in the environment when a
package requires pkgtools/install-sh, but this must be done only after the
'depends barrier' because before that the values are not known yet.

This was a problem if a package a/ depended on pkgtools/install-sh and b/ had
some of its booststrap dependencies not installed yet, because those
dependencies were passed an empty INSTALL variable.