[mk/fetch] Support fetching over cvs, git and svn.
Added FETCH_METHOD package variable, which defaults to 'archive', meaning that
a regular archive is to be fetched from MASTER_SITES.

FETCH_METHOD can be set to 'cvs', 'git' or 'svn' in a package Makefile, meaning
that the distfile is to be fetched from a source repository. The MASTER_SITES
is then interpreted as the URL of this repository, using the following format:
   MASTER_SITES= url[@revision[+module]]
The bits between squared brackets are optional, and can be used to specify a
particular commit and module to be fetched in the repository located at 'url'.
The interpretation of revision and module as well as the url is left to the
underlying fetch tool (cvs, git or svn).

The archive extracted out of the repository is stored in DISTDIR as for other
archives, and its checksum is computed according to the contents of the
archive, ommitting irrelevant metadata information (because metadata will vary
depending on the fetcher's uid, date, etc.). The size of the archive is not
either recorded in distinfo (it will vary for the same reasons).

Of course, fetching over cvs,git,svn must not be mistaken with the 'checkout'
target. Fetch still retrieves an archive from a fixed commit number, while
checkout extracts a complete repository for further manipulation or testing.

This new feature should probably be used sparingly, as making an archive
available so that robotpkg or anyone else can download it is still the
preferred (robust) method.