[archivers/libarchive] Disable lzma support for the inplace version of this pkg
The 'inplace' version of libarchive is statically linked in pkg_install and
digest, so it needs all dependencies properly passed on the link command line.

So make sure libarchive don't detect lzma support because -llzma is not passed
in pkg_install and digest link and is not needed at the moment.