[pkgtools/install-sh] Fix export definitions for INSTALL et al.
Fix in commit d856331fcfb was actually broken in that it always prevented
INSTALL et al. variables to be defined. This is because 'ifeq' conditional are
immediate in gnumake (they evaluate the current value of variables). So instead
use $(if) conditionals  which are expanded only when the variables are used.