[mk/internal] Make the logfilter script also log to a file even in verbose mode
When the logfilter is configured as a no-op (e.g. with VERBOSE=1), it must
still log to the logfile if specified. This is more complex than it seems
because the exit status of the actual command must be extracted from the pipe
in which the 'tee' to the logfile is run.