[graphics/pqp] Fix build target
For some reason, using libtool with regular .o files to produce an ar achive
does not work (the archive is created empty) altough the documentation
explicitely allows it.

Use libtool according to the "standard workflow" instead, and produce both a
shared library and an ar archive. This is the only thing that seem to work with
libtool nowadays.

Bump PKGREVISION.