[mk/depends] Improve error message for missing system dependencies
The new version of the error message for system dependencies was split into a
short summary and a more detailed log file. Apparently, the short summary
should be less short in order not to confuse users.

Also made more visible the printed link to the log file with detailed
instructions.