[mk/sysdep] Fix exportation of AUTOCONF (and other) when not yet resolved
As for some other env variable, AUTOCONF must be defined only when the
dependencies are resolved. Otherwise, it can be set to the wrong /bin/autoconf
when PREFIX.autoconf is not defined yet.

Since this problem is recurrent in depend.mk files, handle this in a more
generic way by 'unexport'ing empty variables at the end of robotpkg.mk. This is
required because exporting an empty variable is different from not exporting a
variable.