[mk] Don't export PREFIX, pass it through ALL_ENV instead
If a package modifies the default PREFIX, this modified value should not be
exported to sub-makes when installing missing dependencies. So instead of
exporting PREFIX, add it to the ALL_ENV variable which is the environment of
all targets but not inherited in sub-makes.

This fixes the build of architecture/openhrp when built as the very first
package in a fresh system.