[mk/update] Do a 'make cleaner' before updating packages
Packages with stale working directories might confuse the 'update'
target, resulting in packages sometimes not reinstalled. Since by default
'update' cleans after itself anyway, it is not a big offense to also clean
beforehand.
The cleaning is done thanks to the new 'cleaner' target, which handles
checkouts correctly (although whether checkouts are correctly rebuilt in this
situation is purely a per-package issue and cannot be guaranteed by robotpkg
itself).