[hardware/platine-genom] fix build on Fedora 14
The following gcc 4.5 change breaks genom pre-processing of
'requires' keywords:
"If a header named in a #include directive is not found, the compiler
exits immediately. This avoids a cascade of errors arising from
declarations expected to be found in that header being missing."

So add explicit -I options to GENOM_ARGS for required packages
that are declared *after* #include'ing one of its file.