[architecture/genom] Fix depend.mk regarding GENOM_ARGS manipualtions
Special care must be taken when manipulating GENOM_ARGS variable so
that its content is not expanded too early (especially if the
GENOM_ARGS variable contains references to PREFIX.<pkg> variables as
in hardware/platine-genom).

The genom options handling in the depend.mk was trying to remove
options from GENOM_ARGS when the corresponding option in robotpkg.conf
was disabled. For doing that, it used ":= style" affectations, which
has bad side effects in this case.

It's quite hard to rewrite this differently, so for now remove all
PKG_OPTION_UNSET.* scripts that are problematic.