[graphics/flann] Import flann-1.7.1
FLANN is a library for performing fast approximate nearest neighbor
searches in high dimensional spaces. It contains a collection of
algorithms we found to work best for nearest neighbor search and a system
for automatically choosing the best algorithm and optimum parameters
depending on the dataset.

FLANN is written in C++ and contains bindings for the following languages:
C, MATLAB and Python.