[graphics/blender] Import blender-2.56
Blender is a suite of tools enabling the creation of and replay of linear and
real-time, interactive 3D content. It offers full functionality for modeling,
rendering, animation, postproduction and game creation and playback with the
singular benefits of cross-platform operability and a download file size of
less than 2.5MB.

Aimed at media professionals and individual creative users, Blender can be
used to create commercials and other broadcast quality linear content, while
the incorporation of a real-time 3D engine allows for the creation of 3D
interactive content for stand-alone playback or integration in a web browser.

Originally developed by the company 'Not a Number' (NaN), Blender now is
continued as 'Free Software', with the sources available under GNU GPL.