[lang/python3] Prefer linking against shared python library instead of static
In the system files search for python, the spec for checking python library was
specifying .a, .so and .dylib in this order. If both static and shared
libraries were present, this would actually find the static library first.

For cmake packages, it looks like the standard python cmake fragment expects to
be given the full path to the library instead of an LDFLAG style directive
(e.g. -lpython). To do this, the file found by the robotpkg system search is
used (thus indirectly prefering the static archive).

This causes problems on some ubuntu x86_64 were the static python archive
appears to be a 32 bit version. To overcome this quickly, just tell robotpkg to
look for the shared library first.