[middleware/ros-core] Import ros-core-1.4.0
Robot Operating System (ROS) is a meta-operating system for your robot.
This package provides the minimal core part of ROS.

A number of patches were necessary, but they are a bit robotpkg specific
because they mostly deal with rpath (Linux) or install_name (Darwin) in order
to cope with the lack of an 'install' target. Properly fixing this would
require the development of a realinstall target. Other patches fix issues with
bash/python path in the scripts.