[bootstrap] Invoke configure scripts via `sh ./configure`
When file mode bits are not set properly, configure scripts may not be
executable. So use the 'sh ./configure' idom to invoke them.

This was actually already done like this, but the shell to use was supposed to
be stored in the 'shprog' variable which was never set by the bootstrap
script (and thus left empty).