[mk/patch] Fix patch ordering with gmake>=3.82
With gmake-3.82, the $(wildcard ) function does not sort its result anymore.
This cannot really be considered as a regression since the documentation does
not mention that the result should be sorted, but since the result of a
wildcard expansion by any shell will be sorted, one could expect the same
behaviour (as it used to be).

Anyway, one must now explicitely sort the patch list so that they are applied
in the correct order.