[archivers/libarchive] Prevent configure script from regenerating itself
Explicitely disable the call to the 'missing' autoconf program (by setting
MISSING to ':'), because the configure script was probably generated in
maintainer mode and insists hard on regenerating itself. Since this fails if
autoconf is not installed, and since autoconf is not a dependency - and should
not be, it's better to avoid running into all this kind of trouble.