[mk/depends] Print a warning for sysdeps whose version cannot be checked
A warning will be printed for system dependencies that have an ABI requirement
but no rule to check the version, since the ABI requirement is useless in this
situation. This also misleads the end user since a message like "found xx>=n"
is printed, although was is really done is "found xx".

Thus, system dependencies whose version cannot be checked (e.g. because the
system tool cannot report it) should define an ABI requirement with no version
restriction.