[mk/depends] Print files that are missing when an installed package is corrupt
Improve the message "a package foo is installed but some files are missing",
printed when a package is recorded in pkgdb but some of the files listed in
SYSTEM_SEARCH are missing. This is either due to a manually hacked PREFIX, or a
typo in the pkg's depend.mk. In both case, printing the path of the missing
files should help finding the source of the problem.