[mk/update] Improve message for packages that cannot be found during an update
When `make update` tries to reinstall an installed package whose category was
changed, it will fail to cd into the original directory. In this case, the
package is simply skipped. Now, it also prints a message after the update
listing those packages that could not be reinstalled.

While here, avoid the 'if cd xxx; then' construct since it apparently fails to
detect non-existent directories with some shells (reported by Thibault Kruse on
private). I could not reproduce the problem with bash or tcsh, but replaced
anyway by 'if test -d xxx; then cd xxx'.