[mk/update] Let 'update' be more robust to category moves or deleted packages
It is not sufficient to test for the existence of a package directory during
'make update'. When a package is moved or deleted, users may have local files
in the original package directory that prevent git pull from deleting the
non-empty directory. robotpkg will then happily cd into this but fail to make
anything because of a missing Makefile (moved or deleted).

To be more robust, one must test the existence of category/pkg/Makefile.