[mk/internal] Don't trust users environment: ignore most of the env variables
Too many people have random environment settings, such a LD_LIBRARY_PATH or
CPATH, that can break the build of packages. As a drastic measure, simply
ignore (reset) any environment variable in robotpkg Makefiles.
If it is necessary for the end-user to pass variables to robotpkg, one must use
the command line or the robotpkg.conf file for that.

The variables currently _not_ ignored are:
  ROBOTPKG_BASE ROBOTPKG_DIR PATH TERM TERMCAP DISPLAY XAUTHORITY SSH_AUTH_SOCK
  MAKEFLAGS MFLAGS MAKEOVERRIDES

This requires a few additional precautions for recursive invocation of ${MAKE}
via the ${RECURSIVE_MAKE} variable so that PKGSRC_MAKE_ENV is passed via the
${MAKE} command line instead of the environment.

Technically, those ignored variables are set to an empty string and unexported,
because gmake-3.81 does not have the 'undefine' directive that was added in
gmake-3.82. This is a problem for constructs like foo?=x, but there are
normally no such constructs involving variables that could come from the
environment.

While here, moved the setting of a C locale from robotpkg.prefs.mk to env.mk
to centralize environment tweaks in env.mk.