[mk/sysdep] Fix (again) cmake version of print-PLIST
The print-PLIST target uses the install_manifest from cmake when a package is
installed via cmake. Commit 4d594fd49 tried to fix the path to this file, but
was not complete because the ${BUILD_DIRS} variable may define either absolute
paths or relative paths (to ${WRKSRC}). To handle both cases, use a $(wildcard)
on the two alternatives: this will filter out the wrong solution...