[mk/depends] Warn if multiple installations of a dependency are found.
Add a check that detects multiple installations of a dependency in the list of
prefixes used by the package. This is usually dangerous since any build command
line will typically have to contain flags pointing to the different prefixes,
making it difficult to properly select the exact dependency.

As a gross heuristic, only prefixes that refer to a SYSTEM_SEARCH.pkg
mentioning an 'include' or 'lib' directory are considered (this filters out
tools like pax, pkg-config etc. that have no header or libraries and which will
not generate any conflicts even if they are found in multiple locations). Along
the same lines, only dependencies that refer to a SYSTEM_SEARCH.pkg containing
'include' or 'lib' are considered.