[pkgtools/pkg_install] Fix 017c6b7deb (deinstall target)
017c6b7deb was wrong and broke the upgrade of pkg_install after a fresh
bootstrap (no pkg_delete tool installed yet).

The correct fix for the deinstall target is to depend on the custom pkg_build
first (as the other update and install targets), so that ${WRKSRC} contains a
valid pkg_delete. To make sure that pkg_build will be the first target of
deinstall, define this before including robotpkg.mk