[mk/depends] Make SYSTEM_SEARCH.<pkg> ELF-centric and improve prefixsearch.sh
PLISTs are ELF-centric, so SYSTEM_SEARCH.<pkg> mimics this as well. This means
that shared libraries entries in a SYSTEM_SEARCH of a dependency must end in
.so (.a are not affected, nor are .la libtool archives)

On systems where _OPSYS_SHLIB_TYPE is set to dylib (Darwin), prefixsearch.sh
replaces all files of the form */lib*.so and */lib*.so.[0-9]* by their
equivalent in .dylib notation and look for them _iff_ the .so is not found.

The actual name of the file found is reported (as before) but for missing
entries all missed attempts are now reported (to avoid generating confusion
with lines like missing xxx.so on Darwin).

As a consequence, removed all explicit .dylib entries from the few depend.mk
and mk/sysdep/*.mk that were mentioning it.