[mk/fetch] Re-enable file verification in the fetch target even for cvs/git/svn
The file verification in the fetch script was disabled for cvs/git/svn
distfiles fetch because even though the checksum is correctly computed, the
expected file size is still unknown.

However, file verification can be reenabled because the distinfo file is now
correctly generated without the file size information (for cvs/git/svn) and the
fetch script doesn't check the size in this situation.

This allows failover to MASTER_BACKUP_SITE for files fetched vian an scm that
have an incorrect checksum.