[graphics/vcglib] Fix fetch target: ignore svn 'external' repositories
vcglib is fetched via a svn repository (no official release), but that
repository contains an external reference on Eigen that is not versionned and
cannot be fetched at a particular version.

Since the vcglib package uses Eigen from the system, it is useless to fetch
this external repository, so just add --ignore-external in FETCH_BEFORE_ARGS.

Regen distinfo.