[mk/depends] Don't blindly add libraries to LDFLAGS
A few old packages where adding their own libraries to the linker flags in
their depend.mk. This is wrong: packages that need this (e.g. pkgtools/digest
to link against libarchive) should set it explictly.

This was inherited from the old times when robotpkg was less stable and had
problems dealing with system vs. robotpkg packages (not that it's totally clean
today... ;).