[graphics/gdhe] Fix compilation glitches on MacOSX
Fix the following issues:
- Pass jpeg prefix to configure
- Define glu include dir
- Add required -install_name flag when building gdhe.dylib (for extensions)
- Do not fail when there is no X11 connection while building pkg_mkIndex

There is still a potential issue with macports, as the later does not provide a
tkInt.h header, so it has to be copied manually to some place where gdhe tries
to find it. So keep the default ONLY_FOR_PLATFORM settings for now.

No need to bump PKGREVISION.