[mk/depends] Raise an error for system packages still installed by robotpkg
The typical scenario for which this check is necessary is when a package
defaults to 'system' (but still available in robotpkg) and a user installs it
manually with robotpkg. Later, other packages using this one as a dependency
will randomly pick up one of the installation (system or robotpkg), leading to
various kind of mysterious, frightening, horrible and error-prone situations.
Gasp.